polymarket_bot_py/requirements.txt
py-clob-client
websocket-client
python-dotenv
aiohttp
